requirements.txt
pandas
requests
pyarrow
rich         
python-dotenv 
fastapi
nfl_data_py
psycopg[binary]